requirements.txt
autopep8==2.0.2
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
python-dotenv==1.0.0
rsa==4.9
Telethon==1.27.0
